1.9 Heap Allocation: followed implementation suggested in tutorial
- Compiler's generated assembly is now only and completely 64-bit.
- Implemented pairs: code generation and C runtime printing.

+++ src/ctest.c
@@ -3,5 +3,5 @@
 ptr scheme_entry(context* ctxt, char* stack_base, char* heap){
   ctxt->ecx = (void*) 2;
   *(stack_base-8) = 7;
-  return *(stack_base-8) << (int) ctxt->ecx;
+  return *(stack_base-8) << (long) ctxt->ecx;
 }
